+++ common/src/main/java/fr/loudo/narrativecraft/network/handlers/ServerPacketHandler.java
@@ -555,6 +555,10 @@ public class ServerPacketHandler {
             storyHandler.stop();
         }
 
+        if (packet.newGame()) {
+            NarrativeCraftMod.getInstance().getSaveFileManager().removeSaveFile((ServerPlayer) player);
+        }
+
         try {
             storyHandler = null;
             if (packet.fromSave()) {